+++ loader.php
@@ -5,12 +5,6 @@
  * @package PuppyFW
  */
 
-if ( defined( 'PUPPYFW_VERSION' ) ) {
-	return;
-}
-
-define( 'PUPPYFW_VERSION', '0.2.0' );
-
 if ( ! defined( 'PUPPYFW_PATH' ) ) {
 	define( 'PUPPYFW_PATH', plugin_dir_path( __FILE__ ) );
 }