+++ puppyfw.php
@@ -12,4 +12,8 @@
  * @package         PuppyFW
  */
 
+if ( function_exists( 'puppyfw' ) ) {
+	return;
+}
+
 require_once plugin_dir_path( __FILE__ ) . 'vendor/autoload.php';